+++ src/Relation/Relationship.php
@@ -6,12 +6,17 @@ use InvalidArgumentException;
 use Pulsar\Model;
 use Pulsar\Property;
 
-class RelationFactory
+class Relationship
 {
+    const HAS_ONE = 'has_one';
+    const HAS_MANY = 'has_many';
+    const BELONGS_TO = 'belongs_to';
+    const BELONGS_TO_MANY = 'belongs_to_many';
+
     /**
      * Creates a new relation instance given a model and property.
      */
-    public static function make(Model $model, string $propertyName, Property $property): Relation
+    public static function make(Model $model, string $propertyName, Property $property): AbstractRelation
     {
         $relationModelClass = $property->getRelation();
         if (!$relationModelClass) {
@@ -22,19 +27,19 @@ class RelationFactory
         $localKey = $property->getLocalKey();
         $relationType = $property->getRelationType();
 
-        if (Model::RELATIONSHIP_HAS_ONE == $relationType) {
+        if (self::HAS_ONE == $relationType) {
             return new HasOne($model, $localKey, $relationModelClass, $foreignKey);
         }
 
-        if (Model::RELATIONSHIP_HAS_MANY == $relationType) {
+        if (self::HAS_MANY == $relationType) {
             return new HasMany($model, $localKey, $relationModelClass, $foreignKey);
         }
 
-        if (Model::RELATIONSHIP_BELONGS_TO == $relationType) {
+        if (self::BELONGS_TO == $relationType) {
             return new BelongsTo($model, $localKey, $relationModelClass, $foreignKey);
         }
 
-        if (Model::RELATIONSHIP_BELONGS_TO_MANY == $relationType) {
+        if (self::BELONGS_TO_MANY == $relationType) {
             $pivotTable = $property->getPivotTablename();
 
             return new BelongsToMany($model, $localKey, $pivotTable, $relationModelClass, $foreignKey);