+++ tests/Models/Balance.php
@@ -3,6 +3,7 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
+use Pulsar\Relation\Relationship;
 use Pulsar\Type;
 
 class Balance extends Model
@@ -10,7 +11,7 @@ class Balance extends Model
     protected static $properties = [
         'person' => [
             'relation' => Person::class,
-            'relation_type' => Model::RELATIONSHIP_BELONGS_TO,
+            'relation_type' => Relationship::BELONGS_TO,
         ],
         'amount' => [
             'type' => Type::FLOAT,