+++ tests/Models/Car.php
@@ -3,6 +3,7 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
+use Pulsar\Relation\Relationship;
 use Pulsar\Type;
 
 class Car extends Model
@@ -12,7 +13,7 @@ class Car extends Model
         'model' => [],
         'garage_id' => [
             'type' => Type::INTEGER,
-            'relation_type' => Model::RELATIONSHIP_BELONGS_TO,
+            'relation_type' => Relationship::BELONGS_TO,
         ],
     ];
 }