+++ tests/Models/Category.php
@@ -3,6 +3,7 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
+use Pulsar\Relation\Relationship;
 
 class Category extends Model
 {
@@ -10,7 +11,7 @@ class Category extends Model
         'name' => [],
         'posts' => [
             'relation' => Post::class,
-            'relation_type' => Model::RELATIONSHIP_HAS_MANY,
+            'relation_type' => Relationship::HAS_MANY,
         ],
     ];
 }