+++ tests/Models/Garage.php
@@ -3,6 +3,7 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
+use Pulsar\Relation\Relationship;
 use Pulsar\Type;
 
 class Garage extends Model
@@ -10,11 +11,11 @@ class Garage extends Model
     protected static $properties = [
         'person_id' => [
             'type' => Type::INTEGER,
-            'relation_type' => Model::RELATIONSHIP_BELONGS_TO,
+            'relation_type' => Relationship::BELONGS_TO,
         ],
         'location' => [],
         'cars' => [
-            'relation_type' => Model::RELATIONSHIP_HAS_MANY,
+            'relation_type' => Relationship::HAS_MANY,
         ],
     ];
 }