+++ tests/Models/Invoice.php
@@ -3,13 +3,14 @@
 namespace Pulsar\Tests\Models;
 
 use Pulsar\Model;
+use Pulsar\Relation\Relationship;
 
 class Invoice extends Model
 {
     public static $properties = [
         'customer' => [
             'relation' => Customer::class,
-            'relation_type' => self::RELATIONSHIP_BELONGS_TO,
+            'relation_type' => Relationship::BELONGS_TO,
         ],
     ];
 }