+++ tests/Models/Person.php
@@ -4,6 +4,7 @@ namespace Pulsar\Tests\Models;
 
 use Pulsar\ACLModel;
 use Pulsar\Model;
+use Pulsar\Relation\Relationship;
 use Pulsar\Type;
 
 class Person extends ACLModel
@@ -22,7 +23,7 @@ class Person extends ACLModel
         ],
         'garage' => [
             'relation' => Garage::class,
-            'relation_type' => Model::RELATIONSHIP_HAS_ONE,
+            'relation_type' => Relationship::HAS_ONE,
         ],
     ];