+++ tests/Relation/RelationTest.php
@@ -13,7 +13,7 @@ class RelationTest extends MockeryTestCase
     public function testGetLocalModel()
     {
         $model = Mockery::mock(Model::class);
-        $relation = new TestRelation($model, 'user_id', TestModel::class, 'id');
+        $relation = new TestAbstractRelation($model, 'user_id', TestModel::class, 'id');
 
         $this->assertEquals($model, $relation->getLocalModel());
     }
@@ -21,7 +21,7 @@ class RelationTest extends MockeryTestCase
     public function testGetLocalKey()
     {
         $model = Mockery::mock(Model::class);
-        $relation = new TestRelation($model, 'user_id', TestModel::class, 'id');
+        $relation = new TestAbstractRelation($model, 'user_id', TestModel::class, 'id');
 
         $this->assertEquals('user_id', $relation->getLocalKey());
     }
@@ -29,7 +29,7 @@ class RelationTest extends MockeryTestCase
     public function testGetForeignModel()
     {
         $model = Mockery::mock(Model::class);
-        $relation = new TestRelation($model, 'user_id', TestModel::class, 'id');
+        $relation = new TestAbstractRelation($model, 'user_id', TestModel::class, 'id');
 
         $this->assertEquals(TestModel::class, $relation->getForeignModel());
     }
@@ -37,7 +37,7 @@ class RelationTest extends MockeryTestCase
     public function testGetForeignKey()
     {
         $model = Mockery::mock(Model::class);
-        $relation = new TestRelation($model, 'user_id', TestModel::class, 'id');
+        $relation = new TestAbstractRelation($model, 'user_id', TestModel::class, 'id');
 
         $this->assertEquals('id', $relation->getForeignKey());
     }
@@ -45,7 +45,7 @@ class RelationTest extends MockeryTestCase
     public function testGetQuery()
     {
         $model = Mockery::mock(Model::class);
-        $relation = new TestRelation($model, 'user_id', TestModel::class, 'id');
+        $relation = new TestAbstractRelation($model, 'user_id', TestModel::class, 'id');
 
         $query = $relation->getQuery();
         $this->assertInstanceOf(Query::class, $query);
@@ -55,7 +55,7 @@ class RelationTest extends MockeryTestCase
     public function testCallOnQuery()
     {
         $model = Mockery::mock(Model::class);
-        $relation = new TestRelation($model, 'user_id', TestModel::class, 'id');
+        $relation = new TestAbstractRelation($model, 'user_id', TestModel::class, 'id');
 
         $query = $relation->where(['name' => 'Bob']);